chore: extract eval
